Permissions: Remove AsaApi function exports
Permissions uses the ARK_EXPORTS and ARK_API defines in the same way as
the AsaApi headers. This results in Permissions.dll exporting functions
from the included AsaApi headers. This change replaces ARK_EXPORTS and
ARK_API in the Permissions code with PERMISSIONS_EXPORTS and
PERMISSIONS_API so that Permissions.dll only exports Permissions
functions.