Permissions: Improve Permissions.dll exports (#11)
* Permissions: Remove AsaApi function exports

Permissions uses the ARK_EXPORTS and ARK_API defines in the same way as
the AsaApi headers. This results in Permissions.dll exporting functions
from the included AsaApi headers. This change replaces ARK_EXPORTS and
ARK_API in the Permissions code with PERMISSIONS_EXPORTS and
PERMISSIONS_API so that Permissions.dll only exports Permissions
functions.

* Permissions: Export AddPlayerPermissionCallback()

AddPlayerPermissionCallback() is declared as an API function in
Permissions.h. However, the declaration does not match the function
definition, resulting in the function not being exported by the DLL.
This change fixes the declaration to match the definition.